IGNITE-28813 After review #1.2
